new GCC = new warning messages, yippeeeecvs commit code/parse/encrypt.h code/parse/encrypt.cpp code/scramble/scramble.cpp

+++ code/math/vecmat.h
@@ -9,13 +9,17 @@
 
 /*
  * $Logfile: /Freespace2/code/Math/VecMat.h $
- * $Revision: 2.8 $
- * $Date: 2005-04-12 05:26:36 $
+ * $Revision: 2.9 $
+ * $Date: 2005-04-15 11:36:54 $
  * $Author: taylor $
  *
  * Header file for functions that manipulate vectors and matricies
  *
  * $Log: not supported by cvs2svn $
+ * Revision 2.8  2005/04/12 05:26:36  taylor
+ * many, many compiler warning and header fixes (Jens Granseuer)
+ * fix free on possible NULL in modelinterp.cpp (Jens Granseuer)
+ *
  * Revision 2.7  2005/04/05 05:53:19  taylor
  * s/vector/vec3d/g, better support for different compilers (Jens Granseuer)
  *
@@ -245,10 +249,10 @@ extern matrix vmd_identity_matrix;
 
 //Here's a handy constant
 
-#define ZERO_VECTOR {(float)0.0,(float)0.0,(float)0.0}
-#define IDENTITY_MATRIX {(float)1.0,(float)0.0,(float)0.0,	\
-                         (float)0.0,(float)1.0,(float)0.0,	\
-                         (float)0.0,(float)0.0,(float)1.0 }
+#define ZERO_VECTOR {0.0f,0.0f,0.0f}
+#define IDENTITY_MATRIX {1.0f,0.0f,0.0f,	\
+                         0.0f,1.0f,0.0f,	\
+                         0.0f,0.0f,1.0f }
 
 //fills in fields of an angle vector
 #define vm_angvec_make(v,_p,_b,_h) (((v)->p=(_p), (v)->b=(_b), (v)->h=(_h)), (v))
